fred_oil_prices_api/requirements.txt
requests
google-cloud-pubsub
fastavro